clk/qcom: Treat different GDSCs didderently
Apparently not all GDSCs are the same. In Linux driver, depending on which
GDSC flags are set, different status register is checked when powering it
on/off. And on top of that, different bit inside that register is tested.

Port missing parts from linux driver to U-Boot:
 - add GDSC flags;
 - adjust logic when checking GDSC status to match one in linux driver.

This fixes e.g. sdm660's USB_30_GDSC to be forever stuck during power on,
since unlike sdm845 one it doesn't have POLL_CFG_GDSCR flag.

Signed-off-by: Alexey Minnekhanov <alexeymin@postmarketos.org>