CI: Add CI to build/test U-Boot for SDM660
Simple build script for PRs.
Also create Android boot images for merge request's CI.

And add auto-rebaser job, which will run rebase on upstream/master
once a day to try to see if there are some incompatibilities with
upstream.

Add release branch handling and automatically publish a release
on github when a tag is pushed.

Signed-off-by: Alexey Minnekhanov <alexeymin@minlexx.ru>